changed CPU time to double from int

+++ simulation.cpp
@@ -115,7 +115,7 @@ int main(int argc, char **argv){
 	//cout << "Solution: " << solution << endl;
 
 	endTime = clock();
-	int totalTime = (int)(((endTime - startTime) / double(CLOCKS_PER_SEC)) * 100);
+	double totalTime = (((endTime - startTime) / double(CLOCKS_PER_SEC)) * 100);
 	//cout << "CPU Time = " << totalTime << " milliseconds.\n";
 	cout << totalTime << endl;
